refactor(app): add folder for constant variables
